Implement CurriculumForm
